alot of refactoring stuffs and restructurign
