+++ rosbot_controller/src/mppi/utils/profiler.py
@@ -2,6 +2,7 @@ from functools import wraps
 import time
 from rospy import loginfo_throttle, loginfo
 
+
 def profile_throttle(tag, freq):
     def inner_f(f):
         @wraps(f)
